# api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/trainings/{trainingId}:
    get:
      tags:
      - training-rest-controller
      summary: Get a training by ID
      operationId: getTrainingByIdd
      parameters:
      - name: trainingId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Training not found
          content: {}
        "200":
          description: Training found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UniqueTrainingDTO"
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "403":
          description: Forbidden - You don't have permission to access
        "500":
          description: Internal server error
    put:
      tags:
      - training-rest-controller
      summary: Replace a training
      operationId: replacePost
      parameters:
      - name: trainingId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingDTO"
        required: true
      responses:
        "404":
          description: Training not found
          content: {}
        "200":
          description: Training replaced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingDTO"
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "403":
          description: Forbidden - You don't have permission to access
        "500":
          description: Internal server error
    delete:
      tags:
      - training-rest-controller
      summary: Delete a training
      operationId: deleteTraining
      parameters:
      - name: trainingId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Training not found
          content: {}
        "200":
          description: Training deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingDTO"
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "403":
          description: Forbidden - You don't have permission to access
        "500":
          description: Internal server error
  /api/trainings/{trainingId}/image:
    get:
      tags:
      - training-rest-controller
      summary: Retrieve training image
      operationId: getTrainingImage
      parameters:
      - name: trainingId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Training image not found
          content: {}
        "200":
          description: Training image found
        "500":
          description: Internal server error
    put:
      tags:
      - training-rest-controller
      summary: Replace training image
      operationId: replaceTrainingImage
      parameters:
      - name: trainingId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "404":
          description: Not Found
        "204":
          description: Training image replaced
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "500":
          description: Internal server error
    post:
      tags:
      - training-rest-controller
      summary: Create training image
      operationId: createTrainingImage
      parameters:
      - name: trainingId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "404":
          description: Not Found
        "201":
          description: Training image created
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "500":
          description: Internal server error
    delete:
      tags:
      - training-rest-controller
      summary: Delete training image
      operationId: deletePostImage
      parameters:
      - name: trainingId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Training image not found
          content: {}
        "204":
          description: Training image deleted
        "500":
          description: Internal server error
...
  /api/trainings/:
    get:
      tags:
      - training-rest-controller
      summary: Get all trainings
      operationId: getTrainings
      responses:
        "404":
          description: Trainings not found
          content: {}
        "200":
          description: Found all trainings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingDTO"
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "403":
          description: Forbidden - You don't have permission to access
        "500":
          description: Internal server error
    post:
      tags:
      - training-rest-controller
      summary: Create a training
      operationId: createTraining
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "201":
          description: Training created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingDTO"
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "403":
          description: Forbidden - You don't have permission to access
        "500":
          description: Internal server error
...
  /api/trainings/paginated:
    get:
      tags:
      - training-rest-controller
      summary: Get paginated trainings
      operationId: getPaginatedTrainings
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "404":
          description: Not Found
        "200":
          description: Trainings retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingDTO"
        "400":
          description: Bad request - Invalid pagination parameters
        "500":
          description: Internal server error
...
components:
  schemas:
...
    TrainingDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        goal:
          type: string
        intensity:
          type: string
...
    UniqueTrainingDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        goal:
          type: string
        intensity:
          type: string